Subject: Click listener for Edit Button
What Happened: Finally got click listener to work for both the graphic and the EditButton. Not sure where the problem was or if the problem will appear again. Learning how to set up double click listeners and will need to play around with this some more. Want to commit this stable appearing for now version.