Subject: Obtain rowID for each row in the database
What Happened: Forgot that I had to add _ID when I did the original query in MainActivity. Updated db query to include DB and then store the rowID in ViewHolder. Finally retrieved rowID value and passed it through onClick as a parameter. Seemed there was an error in the original commit- had to redo.